Makefile supportant TEST + SAVE

+++ compaction.cpp
@@ -3,7 +3,7 @@
 #include <utility>
 
 // assumming that to_reduce are in order
-void compaction(ExperimentalResults& res, const std::vector<int> sizes,
+void compaction(ExperimentalResults& res, const std::vector<long> sizes,
 								const std::vector<ExperimentalResults>& to_reduce) {
 
 	// Sum of the counts
@@ -14,12 +14,12 @@ void compaction(ExperimentalResults& res, const std::vector<int> sizes,
 	}
 
 	// Compaction of the neutrons
-	const int m = to_reduce.size();
-	int i = 0;
-	int j = m-1;
+	const long m = to_reduce.size();
+	long i = 0;
+	long j = m-1;
 	while (i < j) {
-		int i0 = to_reduce[i].b; // free place (>=)
-		int j0 = to_reduce[j].b; // elements to move are (<)
+		long i0 = to_reduce[i].b; // free place (>=)
+		long j0 = to_reduce[j].b; // elements to move are (<)
 
 		while (i0 < sizes[i] && j0 > 0) {
 			std::swap(to_reduce[i].absorbed[i0], to_reduce[j].absorbed[j0]);
