+++ compaction.h
@@ -5,7 +5,7 @@
 
 #include "neutron.h"
 
-void compaction(ExperimentalResults& res, const std::vector<int> sizes,
+void compaction(ExperimentalResults& res, const std::vector<long> sizes,
 								const std::vector<ExperimentalResults>& to_reduce);
 
 #endif // COMPACTION_H
