+++ neutron_gpu_caller.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-ExperimentalResults neutron_gpu_caller(float* absorbed, int n,
+ExperimentalResults neutron_gpu_caller(float* absorbed, long n,
 																			 const ProblemParameters& params,
 																			 const std::vector<unsigned long long>& seeds,
 																			 int threadsPerBlock, int neutronsPerThread);
